docshell/test/test_bug653741.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=653741
-->
<head>
  <title>Test for Bug 653741</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=653741">Mozilla Bug 653741</a>

<script type="application/javascript;version=1.7">

/** Test for Bug 653741 **/
SimpleTest.waitForExplicitFinish();

function childLoad() {
  // Spin the event loop so we leave the onload handler.
  SimpleTest.executeSoon(childLoad2);
}

function childLoad2() {
  let cw = $('iframe').contentWindow;
  
  // Save the Y offset.  For sanity's sake, make sure it's not 0, because we
  // should be at the bottom of the page!
  let origYOffset = cw.pageYOffset;
  ok(origYOffset != 0, 'Original Y offset is not 0.');

  // Scroll the iframe to the top, then navigate to #bottom again.
  cw.scrollTo(0, 0);

  // Our current location is #bottom, so this should scroll us down to the
  // bottom again.
  cw.location = cw.location + '';

  is(cw.pageYOffset, origYOffset, 'Correct offset after reloading page.');
  SimpleTest.finish();
}

</script>

<iframe height='100px' id='iframe' src='file_bug653741.html#bottom'></iframe>

</body>
</html>
